move home page helper functions out of views.py into app/home.py
keeps views.py limited to view functions, matching the existing
pattern of history_processor.py and statistics.py for concern-specific
logic modules.
